libdram: check if board config directory exist before clean
When we call clean_libdram we should check if boards/${MTK_BOARD}
directory exist before run `rm -r`.

+++ build_libdram.sh
@@ -11,10 +11,9 @@ function clean_libdram {
     local LIBDRAM_CONFIG="$3"
 
     [ -d "${MTK_BUILD}" ] && rm -r "${MTK_BUILD}"
-    if [ -e "${LIBDRAM_CONFIG}" ]; then
+    if [ -e "${LIBDRAM_CONFIG}" ] && [ -d "boards/${MTK_BOARD}/" ]; then
         rm -r "boards/${MTK_BOARD}/"
     fi
-
 }
 
 function build_libdram {
